requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
Django==4.1.7
executing==1.2.0
fonttools==4.39.0
greenlet==2.0.2
importlib-resources==5.12.0
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.2
packaging==23.0
pandas==2.0.0rc0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pure-eval==0.2.2
Pygments==2.14.0
pygwalker==0.1.4.9.post0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.5.post1
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
wcwidth==0.2.6
zipp==3.15.0
